Add means to locate components in large circuit
Closes #1314
