+++ workcraft/CircuitPlugin/src/org/workcraft/plugins/circuit/VisualCircuit.java
@@ -7,46 +7,34 @@ import org.workcraft.dom.Container;
 import org.workcraft.dom.Node;
 import org.workcraft.dom.hierarchy.NamespaceHelper;
 import org.workcraft.dom.math.MathConnection;
-import org.workcraft.dom.math.MathModel;
 import org.workcraft.dom.math.MathNode;
-import org.workcraft.dom.references.FileReference;
 import org.workcraft.dom.visual.*;
 import org.workcraft.dom.visual.connections.VisualConnection;
-import org.workcraft.exceptions.DeserialisationException;
 import org.workcraft.exceptions.InvalidConnectionException;
-import org.workcraft.formula.jj.ParseException;
-import org.workcraft.formula.visitors.StringGenerator;
 import org.workcraft.gui.properties.ModelProperties;
-import org.workcraft.gui.properties.PropertyDeclaration;
 import org.workcraft.gui.properties.PropertyDescriptor;
-import org.workcraft.gui.properties.PropertyHelper;
 import org.workcraft.gui.tools.CommentGeneratorTool;
 import org.workcraft.gui.tools.Decorator;
 import org.workcraft.plugins.circuit.commands.CircuitLayoutCommand;
 import org.workcraft.plugins.circuit.commands.CircuitLayoutSettings;
-import org.workcraft.plugins.circuit.refinement.ComponentInterface;
 import org.workcraft.plugins.circuit.routing.RouterClient;
 import org.workcraft.plugins.circuit.routing.RouterVisualiser;
 import org.workcraft.plugins.circuit.routing.impl.Router;
 import org.workcraft.plugins.circuit.routing.impl.RouterTask;
 import org.workcraft.plugins.circuit.tools.*;
 import org.workcraft.plugins.circuit.utils.CircuitUtils;
-import org.workcraft.plugins.circuit.utils.RefinementUtils;
-import org.workcraft.plugins.stg.Stg;
-import org.workcraft.utils.*;
-import org.workcraft.workspace.ModelEntry;
+import org.workcraft.utils.Hierarchy;
 
 import java.awt.*;
 import java.awt.geom.Point2D;
-import java.io.File;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.LinkedList;
 
 @DisplayName("Digital Circuit")
 @ShortName("circuit")
 public class VisualCircuit extends AbstractVisualModel {
-
-    public static final String PROPERTY_ENVIRONMENT = "Environment";
-    private static final String REFINEMENT_MODEL_TITLE = "Refinement model";
 
     public VisualCircuit(Circuit model) {
         this(model, null);
@@ -363,18 +351,19 @@ public class VisualCircuit extends AbstractVisualModel {
     public ModelProperties getProperties(VisualNode node) {
         ModelProperties properties = super.getProperties(node);
         if (node == null) {
-            properties.add(getEnvironmentProperty());
+            properties.add(CircuitPropertyHelper.getEnvironmentProperty(this));
+            properties.addAll(CircuitPropertyHelper.getComponentProperties(this));
         } else if (node instanceof VisualFunctionContact) {
             VisualFunctionContact contact = (VisualFunctionContact) node;
-            properties.add(getSetFunctionProperty(contact));
-            properties.add(getResetFunctionProperty(contact));
+            properties.add(CircuitPropertyHelper.getSetFunctionProperty(this, contact));
+            properties.add(CircuitPropertyHelper.getResetFunctionProperty(this, contact));
         } else if (node instanceof VisualFunctionComponent) {
             VisualFunctionComponent component = (VisualFunctionComponent) node;
-            properties.add(getRefinementProperty(component));
+            properties.add(CircuitPropertyHelper.getRefinementProperty(this, component));
             VisualFunctionContact mainOutput = component.getMainVisualOutput();
             if (mainOutput != null) {
-                properties.add(getSetFunctionProperty(mainOutput));
-                properties.add(getResetFunctionProperty(mainOutput));
+                properties.add(CircuitPropertyHelper.getSetFunctionProperty(this, mainOutput));
+                properties.add(CircuitPropertyHelper.getResetFunctionProperty(this, mainOutput));
                 for (PropertyDescriptor property : mainOutput.getDescriptors()) {
                     String propertyName = property.getName();
                     if (Contact.PROPERTY_INIT_TO_ONE.equals(propertyName)
@@ -386,126 +375,6 @@ public class VisualCircuit extends AbstractVisualModel {
             }
         }
         return properties;
-    }
-
-    private PropertyDescriptor getEnvironmentProperty() {
-        return new PropertyDeclaration<>(FileReference.class, PROPERTY_ENVIRONMENT,
-                getMathModel()::setEnvironment, getMathModel()::getEnvironment);
-    }
-
-    private PropertyDescriptor getRefinementProperty(VisualFunctionComponent component) {
-        return new PropertyDeclaration<>(FileReference.class, CircuitComponent.PROPERTY_REFINEMENT,
-                value -> setRefinementIfCompatible(component, value),
-                () -> component.getReferencedComponent().getRefinement());
-    }
-
-    private void setRefinementIfCompatible(VisualFunctionComponent component, FileReference value) {
-        if (value == null) {
-            component.getReferencedComponent().setRefinement(null);
-            return;
-        }
-
-        MathModel refinementModel;
-        File file = value.getFile();
-        try {
-            ModelEntry me = WorkUtils.loadModel(file);
-            refinementModel = me.getMathModel();
-        } catch (DeserialisationException e) {
-            String path = FileUtils.getFullPath(file);
-            DialogUtils.showError("Cannot read refinement model from '" + path + "':\n " + e.getMessage(),
-                    REFINEMENT_MODEL_TITLE);
-            return;
-        }
-
-        if (!(refinementModel instanceof Stg) && !(refinementModel instanceof Circuit)) {
-            DialogUtils.showError("Incompatible refinement model type: " + refinementModel.getDisplayName(),
-                    REFINEMENT_MODEL_TITLE);
-            return;
-        }
-
-        String refinementTitle = refinementModel.getTitle();
-        String componentLabel = component.getLabel();
-        if (!RefinementUtils.isCompatibleName(refinementTitle, componentLabel)) {
-            int answer = DialogUtils.showYesNoCancel(
-                    "Refinement title does not match component label."
-                            + "\n" + getBulletPair("Refinement title", refinementTitle)
-                            + "\n" + getBulletPair("Component label", componentLabel)
-                            + "\n\nUpdate component label to match refinement title?",
-                    REFINEMENT_MODEL_TITLE, 0);
-
-            if (answer == 2) {
-                return;
-            }
-            if (answer == 0) {
-                component.getReferencedComponent().setModule(refinementTitle);
-            }
-        }
-
-        ComponentInterface refinementInterface = RefinementUtils.getModelInterface(refinementModel);
-        ComponentInterface componentInterface = RefinementUtils.getComponentInterface(component.getReferencedComponent());
-        if (!RefinementUtils.isCompatibleSignals(refinementInterface, componentInterface)) {
-            Set<String> missingPins = componentInterface.getMissingSignals(refinementInterface);
-            Set<String> extraPins = componentInterface.getExtraSignals(refinementInterface);
-            Set<String> mismatchPins = componentInterface.getMismatchSignals(refinementInterface);
-            int answer = DialogUtils.showYesNoCancel("Refinement interface signals do not match component pins."
-                    + getBulletPair("Missing component pin", missingPins)
-                    + getBulletPair("Unexpected component pin", extraPins)
-                    + getBulletPair("Incorrect I/O type pin", mismatchPins)
-                    + "\n\nUpdate component pins to match refinement model signals?",
-                    REFINEMENT_MODEL_TITLE, 0);
-            if (answer == 2) {
-                return;
-            }
-            if (answer == 0) {
-                RefinementUtils.updateInterface(this, component, refinementInterface);
-            }
-        }
-        component.getReferencedComponent().setRefinement(value);
-    }
-
-    private String getBulletPair(String key, String value) {
-        return PropertyHelper.BULLET_PREFIX + key + ((value == null) || value.isEmpty() ? " is empty" : (": " + value));
-    }
-
-    private String getBulletPair(String key, Set<String> value) {
-        if (value.isEmpty()) {
-            return "";
-        }
-        return "\n" + PropertyHelper.BULLET_PREFIX + key + (value.size() > 1 ? "s: " : ": ") + TextUtils.wrapItems(value);
-    }
-
-    private PropertyDescriptor getSetFunctionProperty(VisualFunctionContact contact) {
-        return new PropertyDeclaration<String>(String.class, FunctionContact.PROPERTY_SET_FUNCTION,
-                value -> {
-                    try {
-                        contact.setSetFunction(CircuitUtils.parseContactFunction(this, contact, value));
-                    } catch (ParseException e) {
-                        throw new RuntimeException(e);
-                    }
-                },
-                () -> StringGenerator.toString(contact.getSetFunction())) {
-            @Override
-            public boolean isVisible() {
-                return contact.isDriver();
-            }
-        }.setCombinable();
-    }
-
-    private PropertyDescriptor getResetFunctionProperty(VisualFunctionContact contact) {
-        return new PropertyDeclaration<String>(String.class, FunctionContact.PROPERTY_RESET_FUNCTION,
-                value -> {
-                    try {
-                        contact.setResetFunction(CircuitUtils.parseContactFunction(this, contact, value));
-                    } catch (ParseException e) {
-                        throw new RuntimeException(e);
-                    }
-                },
-                () -> StringGenerator.toString(contact.getResetFunction())) {
-            @Override
-            public boolean isVisible() {
-                return contact.isDriver();
-            }
-        }.setCombinable();
     }
 
 }
